+++ LL/src/Core/Application/UseCases/Characters/Events/CharacterLevelUpEvent.cs
@@ -4,5 +4,5 @@ namespace Application.UseCases.Characters.Events;
 public record CharacterLevelUpEvent(
     Guid CharacterId,
     int Level,
-    float Experience,
-    float ExperienceUntilNextLevel) : INotification;
+    long Experience,
+    long ExperienceUntilNextLevel) : INotification;
