+++ LL/src/Core/Application/WebSockets/Contracts/CharacterLevelUpMsg.cs
@@ -3,5 +3,5 @@ namespace Application.WebSockets.Contracts;
 public record CharacterLevelUpMsg(
     Guid CharacterId,
     int Level,
-    float Experience,
-    float ExperienceUntilNextLevel) : GameEventMsg;
+    long Experience,
+    long ExperienceUntilNextLevel) : GameEventMsg;
