+++ LL/src/Core/Domain/Models/Leaderboards/LeaderboardEntry.cs
@@ -5,5 +5,5 @@ public class LeaderboardEntry
     public string CharacterName { get; set; } = string.Empty;
     public int Rank { get; set; }
     public int Level { get; set; }
-    public int Experience { get; set; }
+    public long Experience { get; set; }
 }
